+++ regions.js
@@ -13,25 +13,18 @@ module.exports = {
   getRegion: getRegion,
   setRegion: setRegion,
   getDefaultRegion: getDefaultRegion,
-  //getRegionOrDefault: getRegionOrDefault,
   getInvokedFunctionArnRegion: getInvokedFunctionArnRegion,
   getEventAwsRegions: getEventAwsRegions,
   getEventSourceArnRegions: getEventSourceArnRegions,
-  /** @deprecated simply use `getRegion` directly instead when the region is required */
   configureRegion: configureRegion,
-  //resolveRegion: resolveRegion,
 
   getRegionKey: getRegionKey,
 
-  ONLY_FOR_TESTING: {
-    getRegionRaw: getRegionRaw,
-    getDefaultRegionRaw: getDefaultRegionRaw,
-    setRegionIfNotSet: setRegionIfNotSet
-  }
+  getRegionRaw: getRegionRaw,
+  getDefaultRegionRaw: getDefaultRegionRaw
 };
 
 const Strings = require('core-functions/strings');
-const isBlank = Strings.isBlank;
 const isNotBlank = Strings.isNotBlank;
 const trim = Strings.trim;
 const trimOrEmpty = Strings.trimOrEmpty;
@@ -41,30 +34,12 @@ const getArnRegion = arns.getArnRegion;
 
 /**
  * Gets the region in which this function is running from the `AWS_REGION` environment variable and returns it as is if
- * it's neither "undefined" nor "null"; otherwise logs a warning and returns undefined if it's "undefined" or "null" (or
- * undefined).
- *
- * The `AWS_REGION` environment variable is the best option to use to get the region within AWS Lambda code (and the
- * only option to use at module-level scope), since these environment variables will be set by AWS Lambda.
- *
- * An optional "hidden" 'failFast' boolean argument, which defaults to false, can be passed as true to raise an error if
- * the AWS_REGION env variable is not available or unusable
- *
+ * it's neither "undefined" nor "null"; otherwise returns undefined.
  * @returns {string|undefined} the AWS region (if it exists); otherwise undefined.
  */
 function getRegion() {
   const awsRegion = trim(process.env.AWS_REGION);
-  const region = awsRegion !== "undefined" && awsRegion !== "null" ? awsRegion : undefined;
-  if (!region) {
-    const errorMsg = `Failed to get usable region from AWS_REGION env variable (${awsRegion}) - for unit testing call setRegion beforehand`;
-    const failFast = arguments.length > 0 && arguments[0] === true;
-    if (failFast) {
-      console.error(errorMsg);
-      throw new Error(errorMsg);
-    }
-    console.warn(errorMsg);
-  }
-  return region;
+  return awsRegion !== "undefined" && awsRegion !== "null" ? awsRegion : undefined;
 }
 
 /**
@@ -79,7 +54,7 @@ function setRegion(awsRegion) {
   const region = trim(awsRegion);
   if (region === undefined || region === null || region === 'undefined' || region === 'null') {
     // If the given region is undefined or null, then must delete process.env.AWS_REGION rather than setting it to
-    // undefined or null, which incorrectly sets it to the strings "undefined" or "null" respectively
+    // undefined or null, which sets it to the strings "undefined" or "null" respectively
     delete process.env.AWS_REGION;
   } else {
     process.env.AWS_REGION = region;
@@ -92,27 +67,6 @@ function setRegion(awsRegion) {
  */
 function getDefaultRegion() {
   return trimOrEmpty(process.env.AWS_DEFAULT_REGION);
-}
-
-/**
- * Sets the process.env.AWS_REGION environment variable to the given region, but ONLY if is is not already set!
- * NB: This only sets the region temporarily on the current process and should probably only be used for testing purposes.
- * @deprecated use `setRegion` instead
- * @returns {boolean} true if set; false otherwise.
- */
-function setRegionIfNotSet(awsRegion) {
-  const newRegion = trim(awsRegion);
-
-  // Check if AWS region is already set or not
-  const region = getRegionRaw();
-  if (isBlank(region)) {
-    setRegion(newRegion);
-    return true;
-  }
-  if (process.env.AWS_REGION !== newRegion) {
-    console.log(`Ignoring attempt to change ALREADY set AWS_REGION env variable (${process.env.AWS_REGION}) to (${newRegion})`);
-  }
-  return false;
 }
 
 /**
@@ -163,7 +117,7 @@ function getEventSourceArnRegions(event) {
  * @returns {string|undefined|null} the AWS region (if it exists); otherwise an empty string, undefined or null.
  */
 function getRegionRaw() {
-  return trim(process.env.AWS_REGION);
+  return process.env.AWS_REGION;
 }
 
 /**
@@ -176,38 +130,31 @@ function getRegionRaw() {
  * @returns {string|undefined|null} the AWS default region (if it exists); otherwise an empty string, undefined or null
  */
 function getDefaultRegionRaw() {
-  return trim(process.env.AWS_DEFAULT_REGION);
+  return process.env.AWS_DEFAULT_REGION;
 }
 
 /**
  * Keeps context.region as is if it's already configured on the given context, otherwise gets the current region from
- * process.env.AWS_REGION and if it's not blank, sets it on the context as context.region; otherwise either throws an
- * error if failFast is explicitly true.
- * @deprecated simply use `getRegion` directly instead when the region is required
+ * process.env.AWS_REGION and sets it on the context as context.region.
  * @param {Object|RegionAware} context - a context on which to set the region
- * @param {boolean|undefined} [failFast] - an optional flag that is only used when AWS_REGION is needed and blank and
- * that determines whether the error will be raised (if failFast is explicitly true) or simply logged as a warning
  * @returns {RegionAware} the context with its existing region or the current AWS_REGION env variable value.
- * @throws {Error} if failFast is explicitly true and an AWS_REGION env variable is needed and not available
  */
-function configureRegion(context, failFast) {
+function configureRegion(context) {
   // Resolve the AWS region, if it is not already defined on the context
   if (!context.region) {
-    context.region = getRegion(failFast === true);
+    context.region = getRegion();
   }
-  (context.info || console.log)(`Using region (${getRegion()}) & context.region (${context.region})`);
+  (context.debug || console.log)(`Using region (${process.env.AWS_REGION}) & context.region (${context.region})`);
   return context;
 }
 
 /**
  * Returns the region key object for the given region name.
  * @param {string|undefined} [region] - the name of the region (defaults to current region if not defined)
- * @param {boolean|undefined} [failFast] - an optional flag that is only used when AWS_REGION is needed and blank and
- * that determines whether the error will be raised (if failFast is explicitly true) or simply logged as a warning
  * @returns {{region: string}} a region key object
  */
-function getRegionKey(region, failFast) {
-  const regionName = region ? region : getRegion(failFast === true);
+function getRegionKey(region) {
+  const regionName = region ? region : getRegion();
   let regionKey = regionKeysByRegion.get(regionName);
   if (!regionKey) {
     regionKey = {region: regionName};