+++ test/regions.test.js
@@ -12,8 +12,7 @@ const regions = require('../regions');
 const getRegion = regions.getRegion;
 const setRegion = regions.setRegion;
 
-const getRegionRaw = regions.ONLY_FOR_TESTING.getRegionRaw;
-const setRegionIfNotSet = regions.ONLY_FOR_TESTING.setRegionIfNotSet;
+const getRegionRaw = regions.getRegionRaw;
 
 const Strings = require('core-functions/strings');
 const isBlank = Strings.isBlank;
@@ -44,7 +43,6 @@ test('getRegion and setRegion', t => {
 
     // check get when not set
     t.equal(getRegion(), undefined, `getRegion() must be '${undefined}'`);
-    t.throws(() => getRegion(true), undefined, `getRegion(true) must throw'`);
 
     // check will set, when not set
     const expected = 'TEST_REGION_1';
@@ -57,59 +55,6 @@ test('getRegion and setRegion', t => {
     setRegion(expected2);
     t.equal(getRegion(), expected2, `setRegion('TEST_REGION_3') then getRegion() must now be ${expected2}`);
     t.equal(process.env.AWS_REGION, expected2, `process.env.AWS_REGION must be ${expected2}`);
-
-  } finally {
-    // "Restore" original aws region
-    setRegion(origRegion);
-    // Check "restore" worked
-    if (origRegion === undefined) {
-      t.equal(getRegion(), undefined, `getRegion() must be "restored" to undefined' (orig was ${origRegion})`);
-    } else if (isBlank(origRegion)) {
-      t.equal(getRegion(), '', `getRegion() must be "restored" to empty string (orig was '${origRegion}')`);
-    } else {
-      t.equal(getRegion(), origRegion, `getRegion() must be restored to ${origRegion}`);
-    }
-    t.end();
-  }
-});
-
-// =====================================================================================================================
-// Tests for getRegion and setRegionIfNotSet
-// =====================================================================================================================
-
-test('getRegion and setRegionIfNotSet', t => {
-
-  // Attempt to preserve the original AWS_REGION setting (unfortunately cannot preserve undefined or null)
-  const origRegion = getRegionRaw();
-
-  // check orig
-  if (origRegion === undefined) {
-    t.equal(origRegion, process.env.AWS_REGION, `original raw must be '${process.env.AWS_REGION}'`);
-    t.equal(getRegion(), undefined, `original must be empty string '${process.env.AWS_REGION}'`);
-  } else if (isBlank(origRegion)) {
-    t.equal(origRegion, process.env.AWS_REGION, `original raw must be '${process.env.AWS_REGION}'`);
-    t.equal(getRegion(), '', `original must be undefined '${process.env.AWS_REGION}'`);
-  }
-
-  // Must use delete to "clear" property.env variable - since setting to undefined & null don't work as intended
-  try {
-    // Clear AWS_REGION to undefined (by deleting it)
-    delete process.env.AWS_REGION;
-    t.equal(process.env.AWS_REGION, undefined, `process.env.AWS_REGION must be '${undefined}' after delete`);
-
-    // check get when not set
-    t.equal(getRegion(), undefined, `getRegion() must be '${undefined}'`);
-    t.throws(() => getRegion(true), undefined, `getRegion(true) must throw'`);
-
-    // check will set, when not set
-    const expected = 'TEST_REGION_1';
-    t.ok(setRegionIfNotSet(expected), `setRegionIfNotSet('TEST_REGION_1') must set successfully`);
-    t.equal(getRegion(), expected, `getRegion() must be ${expected}`);
-    t.equal(process.env.AWS_REGION, expected, `process.env.AWS_REGION must be ${expected}`);
-
-    // check was NOT set, when already set set
-    t.notOk(setRegionIfNotSet('TEST_REGION_3'), `setRegionIfNotSet('TEST_REGION_3') must NOT set successfully`);
-    t.equal(getRegion(), expected, `getRegion() must still be ${expected}`);
 
   } finally {
     // "Restore" original aws region